feat: refactor dashboard UI frame
